get_categories updated
Plan for next commit: we're going to want most the regex and logic that's currently in the get_categories and get_items functions to be put into the class definitions instead. Those functions will instead be relegated to iterating through a list of data and creating the objects.
For example, get_categories will simply get a list of top-level URLs and then feed those into Categories objects which will then be added to a list. The class will have all the functions necessary to create its various attributes.

Right now, I have it so that the function makes those attributes and just feeds them into objects. This is not really object-oriented programming at all.